Moved zoo routes to own file and added zoosModel.js file. Still need to move bear routes to separate file.
