+++ index.js
@@ -2,104 +2,22 @@ const express = require('express');
 const helmet = require('helmet');
 const knex = require('knex');
 
+const zoosRoutes = require('./zoos/zoosRoutes.js');
+
 const knexConfig = require('./knexfile');
 
 const db = knex(knexConfig.development);
 
 const server = express();
 
-server.use(express.json());
 server.use(helmet());
+server.use(express.json());
 
 // endpoints here
 ////////////////
 // ZOOS
 ////////////////
-server.get('/api/zoos', (req, res) => {
-  db('zoos')
-    .then((zoos) => {
-      res.status(200).json(zoos);
-    })
-    .catch((err) => {
-      res.status(500).json(err);
-    });
-});
-
-server.get('/api/zoos/:id', (req, res) => {
-  const { id } = req.params;
-  db('zoos')
-    .where({ id })
-    .first()
-    .then((zoo) => {
-      if (!zoo) {
-        return res.status(404).send({
-          message: `The zoo with the specified ID ${id} does not exist.`,
-        });
-      }
-      res.status(200).json(zoo);
-    })
-    .catch((err) => {
-      res.status(500).json(err);
-    });
-});
-
-server.post('/api/zoos', (req, res) => {
-  const { name } = req.body;
-  const zoo = { name };
-  if (!name) {
-    return res.status(400).send({
-      errorMessage: 'Please provide a name for the zoo.',
-    });
-  }
-  db.insert(zoo)
-    .into('zoos')
-    .then((ids) => {
-      res.status(201).json(ids);
-    })
-    .catch((err) => {
-      res.status(500).json(err);
-    });
-});
-
-server.put('/api/zoos/:id', (req, res) => {
-  const { id } = req.params;
-  const { name } = req.body;
-  const newZoo = { name };
-  if (!name) {
-    return res.status(400).send({
-      errorMessage: 'Please provide a name for the zoo.',
-    });
-  }
-  db('zoos')
-    .where({ id })
-    .update(newZoo)
-    .then((count) => {
-      if (!count) {
-        res.status(404).json({
-          message: 'No zoo found to update',
-        });
-      }
-      res.status(200).json({ message: `${count} record updated` });
-    });
-});
-
-server.delete('/api/zoos/:id', (req, res) => {
-  const { id } = req.params;
-  db('zoos')
-    .where({ id })
-    .del()
-    .then((remove) => {
-      if (!remove) {
-        return res.status(404).send({
-          message: `The zoo with the specified ID ${id} does not exist.`,
-        });
-      }
-      res.status(200).send({ message: `zoo with ID ${id} was removed.` });
-    })
-    .catch((err) => {
-      res.status(500).json(err);
-    });
-});
+server.use('/api/zoos', zoosRoutes);
 
 ////////////////
 // BEARS
